fix(snap): create VERSION file during snap build
The makefile requires a `VERSION` file to be present when building the device service,
as it uses `-ldflags` to set the Version value based on the contents of that file.

When building a snap from a clean build, the `VERSION` file is not present.

This commit creates the `VERSION` file before `make build`, using the current git tag
or alternatively falling back to using 0.0.0.

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>